+++ prog/modes/modes.F90
@@ -10,7 +10,7 @@
 !> Program for calculating system normal modes from a Hessian
 program modes
   use dftbp_assert
-  use dftbp_io
+  use dftbp_globalenv, only : stdOut
   use dftbp_initmodes
   use dftbp_accuracy, only : dp, lc
   use dftbp_constants, only : Hartree__cm, Bohr__AA, pi